Refs #1365 hoist plan presence gate
